consolidate inconsistent rules management hook names
