+++ x-pack/plugins/security_solution/public/detection_engine/rule_management/api/hooks/use_fetch_rules_snooze_settings_query.ts
@@ -23,7 +23,7 @@ const FETCH_RULE_SNOOZE_SETTINGS_QUERY_KEY = ['GET', INTERNAL_ALERTING_API_FIND_
  * @param queryOptions - react-query options
  * @returns useQuery result
  */
-export const useFetchRulesSnoozeSettings = (
+export const useFetchRulesSnoozeSettingsQuery = (
   ids: string[],
   queryOptions?: UseQueryOptions<RulesSnoozeSettingsMap, Error, RulesSnoozeSettingsMap, string[]>
 ) => {
